+++ onb/usbonbinterface.cpp
@@ -21,11 +21,11 @@ bool UsbHidOnbInterface::write(Objnet::CommonMessage &msg)
 {
     if (!usb->isReady()) //isOpen())
         return false;
-    unsigned long id = msg.rawId();
+    uint32_t id = msg.rawId();
     unsigned char sz = msg.data().size();
     QByteArray ba;
     ba.resize(13);
-    *reinterpret_cast<unsigned long*>(ba.data()) = id;
+    *reinterpret_cast<uint32_t*>(ba.data()) = id;
     ba[4] = sz;
     for (int i=0; i<sz; i++)
         ba[5+i] = msg.data()[i];
@@ -43,7 +43,7 @@ bool UsbHidOnbInterface::read(Objnet::CommonMessage &msg)
     bool success = usb->readFeature(ba);
     if (success)
     {
-        unsigned long id = *reinterpret_cast<unsigned long*>(ba.data());
+        uint32_t id = *reinterpret_cast<uint32_t*>(ba.data());
         msg.setId(id);
         unsigned char sz = ba[4];
         msg.setData(QByteArray(ba.data() + 5, sz));
@@ -65,7 +65,7 @@ void UsbHidOnbInterface::flush()
     qDebug() << "[UsbHidOnbInterface]: flush is not implemented";
 }
 
-int UsbHidOnbInterface::addFilter(unsigned long id, unsigned long mask)
+int UsbHidOnbInterface::addFilter(uint32_t id, uint32_t mask)
 {
     qDebug() << "[UsbHidOnbInterface]: Filter is not implemented. id=" << id << "mask=" << mask;
     return 0;
